backend-docker/requirements.txt
django==2.0.9
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
django-cors-headers==2.4.0
gunicorn
